+++ docs/pages/visual-test/IconButton-dark.js
@@ -11,6 +11,7 @@ export default function Snapshot(): Node {
           icon="visit"
           bgColor="lightGray"
           iconColor="darkGray"
+          size="md"
         />
       </Box>
     </ColorSchemeProvider>