+++ playwright/accessibility/Tabs_iOS.spec.mjs
@@ -2,7 +2,7 @@
 import { test } from '@playwright/test';
 import expectAccessiblePage from './expectAccessiblePage.mjs';
 
-test('Avatar Accessibility check', async ({ page }) => {
-  await page.goto('/web/avatar');
+test('iOS Tabs Accessibility check', async ({ page }) => {
+  await page.goto('/ios/tabs');
   await expectAccessiblePage({ page });
 });